requirements.txt
xlrd
pandas
xmltodict
Appium-Python-Client
lxml==3.7.2
selenium==3.0.2